Wave 8: 138 functions matched, 438 -> 631, and a promote.py source corruption
Eleven agents over 140 scheduled functions, batched by address locality so
each worked one struct or template family rather than a dozen unrelated
bodies. 138 matched, 1 correctly reclassified as hand-written assembly,
1 parked. Both builds still reproduce SHA1 14dd0b22.

The m4a span is not what data/asm-resident.json assumed. MPlayFadeOut
(sub_080703D4) and m4aMPlayFadeOutPause (sub_08070620) both match as
ordinary C, so ~0x0806F734-0x080718E4 is m4a.c interleaved with m4a_asm.s,
not a monolithic lift. Those two also pin the MusicPlayerTrack offsets.
sub_0806FC50 went the other way: it takes the track pointer in r1 and r0 is
never read, so it is ply_pend, and it names patternLevel (+2),
patternStack[] (+0x44) and cmdPtr (+0x40) in one body. 24 excluded now.

promote.py was silently corrupting the C it emitted. decl_chunks() sniffed
comment boundaries from the first characters of a line instead of tracking
comment state, and two defects compounded: a comment whose prose ended in
`;` closed the chunk mid-comment, and the continuation lines that followed
were then dropped as stray comments because nothing was accumulated any
more. c_0806978C.c was written with an unterminated comment swallowing the
struct beneath it. That one failed loudly; 80 of 117 files this wave were
affected, and the rest of the damage was silent -- deleted documentation and
`#include` lines glued onto the declaration below them, since a directive
could not terminate its own chunk either.

Fixing that exposed two more, both from the same root cause of never having
parsed comments. Dedup keyed on chunk text including the comment, so one
struct documented two ways became `redefinition of struct Mtx43`; the key is
now comment-free while the text kept is still the first occurrence's. And
split_source cut at the definition line, stranding each function's doc
comment in the declaration block -- c_08063F98.c explained how to copy a
matrix directly above the function that loads the identity.

sub_0808AD6C is parked at 85.7%, size-exact. The ROM loads the sub_0808AD68
pool word twice, once as the eors destination which clobbers it and again
for the subtraction; agbcc's CSE keeps a register copy instead. 23 spellings
across 5 probes plus a permuter run did not move it. The note in work/ says
what was tried and why the axis is wrong rather than exhausted.

hardware.h gains DMA3, REG_IE/IF/IME and the interrupt flags. Reaching the
three DMA3 control words through a vu32 cursor is what the ROM has; three
separately named lvalues emit an add between stores despite the addresses
being compile-time constants.

+++ src/decomp/c_08042E18.c
@@ -7,7 +7,13 @@
  * sub_08042E18 @ 0x08042E18
  */
 
-u8 sub_08042E18(int a)
+/* `int`, not the `u8` this was promoted with. Its own bytes cannot tell --
+ * it is a leaf ending in `bx lr` and `ldrb` zero-extends either way -- but
+ * sub_08042DFC forwards the result straight out with no `lsls #24; lsrs #24`,
+ * and a u8-returning callee makes the caller re-narrow. The call site is the
+ * only evidence there is, and it says 32 bits.
+ */
+int sub_08042E18(int a)
 {
     return gUnknown_085D3DD0[a].unk16;
 }
